refactor(daemon): move the run analytics lifecycle to its owning domains
Review follow-up. `createRunAnalyticsLifecycle` takes no Express request or
response, so by `apps/daemon/AGENTS.md` it is a reusable daemon service and
belongs under `src/services/`, not `src/routes/`. The run/project record shapes
it shares with the run routes describe a Run, not an HTTP exchange, so they move
to the runs domain.

The seam also carried a second copy of the analytics facts, with its context,
snapshot and rollout fields widened to `unknown` so the two definitions could
drift. `RunAnalyticsFacts` is now declared once by the lifecycle and reused by
`InternalRunCreationService.start`. Consolidating immediately surfaced what the
weak copy had been hiding: the route was passing a snapshot shape the lifecycle
did not actually accept, and the test was passing an analytics context missing
four required fields. Both are now typed at the choke point.
