+++ apps/daemon/src/runtimes/chat-run-records.ts
@@ -23,7 +23,7 @@ import type {
 import type { RunArtifactDiff } from '../run-artifact-fs.js';
 import type { RunEventForDiagnostics } from '../run-diagnostics.js';
 import type { RunEventForFailureClassification } from '../run-failure-classification.js';
-import type { RunWorkspaceScope } from '../runtimes/project-amr-trace-env.js';
+import type { RunWorkspaceScope } from './project-amr-trace-env.js';
 import type { OdNextRolloutDecision } from '../strategies/od-next/rollout.js';
 import type { OdNextTaskInputSnapshotDescriptor } from '../strategies/od-next/task-input-snapshot.js';
